Add setup of config to create_core
