+++ src/tests/functional/plugin/conformance/test_runner/api_conformance_runner/include/ov_api_conformance_helpers.hpp
@@ -7,6 +7,8 @@
 #include "conformance.hpp"
 #include "common_test_utils/test_constants.hpp"
 
+using namespace ov::test::utils;
+
 namespace ov {
 namespace test {
 namespace conformance {
@@ -14,7 +16,7 @@ namespace conformance {
 inline const std::vector<ov::AnyMap> generate_ov_configs(const std::vector<ov::AnyMap>& config = {}) {
     std::vector<ov::AnyMap> resultConfig;
     for (auto configItem : config) {
-        configItem.insert(ov::test::conformance::pluginConfig.begin(), ov::test::conformance::pluginConfig.end());
+        configItem.insert(ov::test::utils::pluginConfig.begin(), ov::test::utils::pluginConfig.end());
         resultConfig.push_back(configItem);
     }
     return resultConfig;
