+++ src/tests/functional/plugin/conformance/test_runner/api_conformance_runner/src/ov_executable_network/properties.cpp
@@ -18,7 +18,7 @@ const std::vector<ov::AnyMap> inproperties = {
 INSTANTIATE_TEST_SUITE_P(ov_compiled_model_mandatory, OVClassCompiledModelPropertiesIncorrectTests,
                         ::testing::Combine(
                                 ::testing::Values(targetDevice),
-                                ::testing::ValuesIn(generate_ov_configs(inproperties))),
+                                ::testing::ValuesIn(inproperties)),
                         OVClassCompiledModelPropertiesIncorrectTests::getTestCaseName);
 
 const std::vector<ov::AnyMap> default_properties = {
