+++ src/tests/functional/plugin/conformance/test_runner/api_conformance_runner/src/ov_infer_request/callback.cpp
@@ -15,7 +15,7 @@ using namespace ov::test::conformance;
 INSTANTIATE_TEST_SUITE_P(ov_infer_request_mandatory, OVInferRequestCallbackTests,
         ::testing::Combine(
             ::testing::Values(targetDevice),
-            ::testing::Values(pluginConfig)),
+            ::testing::Values(ov::AnyMap({}))),
         OVInferRequestCallbackTests::getTestCaseName);
 
 }  // namespace
