+++ src/tests/functional/plugin/conformance/test_runner/api_conformance_runner/src/ov_infer_request/cancellation.cpp
@@ -12,6 +12,6 @@ using namespace ov::test::conformance;
 INSTANTIATE_TEST_SUITE_P(ov_infer_request_mandatory, OVInferRequestCancellationTests,
         ::testing::Combine(
             ::testing::Values(targetDevice),
-            ::testing::Values(pluginConfig)),
+            ::testing::Values(ov::AnyMap({}))),
         OVInferRequestCancellationTests::getTestCaseName);
 }  // namespace
