+++ src/tests/functional/plugin/conformance/test_runner/api_conformance_runner/src/ov_infer_request/inference_chaining.cpp
@@ -13,12 +13,12 @@ using namespace ov::test::conformance;
 INSTANTIATE_TEST_SUITE_P(ov_infer_request_mandatory, OVInferenceChainingStatic,
                         ::testing::Combine(
                                 ::testing::Values(targetDevice),
-                                ::testing::Values(pluginConfig)),
+                                ::testing::Values(ov::AnyMap({}))),
                         OVInferenceChaining::getTestCaseName);
 
 INSTANTIATE_TEST_SUITE_P(ov_infer_request, OVInferenceChaining,
                         ::testing::Combine(
                                 ::testing::Values(targetDevice),
-                                ::testing::Values(pluginConfig)),
+                                ::testing::Values(ov::AnyMap({}))),
                         OVInferenceChaining::getTestCaseName);
 }  // namespace
