+++ src/tests/functional/plugin/conformance/test_runner/api_conformance_runner/src/ov_infer_request/perf_counters.cpp
@@ -13,7 +13,7 @@ namespace {
 INSTANTIATE_TEST_SUITE_P(ov_infer_request_mandatory, OVInferRequestPerfCountersTest,
                         ::testing::Combine(
                                 ::testing::Values(targetDevice),
-                                ::testing::Values(pluginConfig)),
+                                ::testing::Values(ov::AnyMap({}))),
                          OVInferRequestPerfCountersTest::getTestCaseName);
 
 }  // namespace
