+++ src/tests/functional/plugin/conformance/test_runner/api_conformance_runner/src/ov_infer_request/wait.cpp
@@ -16,7 +16,7 @@ namespace {
 INSTANTIATE_TEST_SUITE_P(ov_infer_request_mandatory, OVInferRequestWaitTests,
                         ::testing::Combine(
                                 ::testing::Values(targetDevice),
-                                ::testing::Values(pluginConfig)),
+                                ::testing::Values(ov::AnyMap({}))),
                             OVInferRequestWaitTests::getTestCaseName);
 
 }  // namespace
