+++ src/tests/functional/plugin/conformance/test_runner/api_conformance_runner/src/ov_plugin/caching_tests.cpp
@@ -42,7 +42,7 @@ INSTANTIATE_TEST_SUITE_P(ov_plugin, CompileModelCacheTestBase,
                                  ::testing::ValuesIn(ovElemTypesTemplate),
                                  ::testing::ValuesIn(ovBatchSizesTemplate),
                                  ::testing::Values(targetDevice),
-                                 ::testing::Values(pluginConfig)),
+                                 ::testing::Values(ov::AnyMap({}))),
                          CompileModelCacheTestBase::getTestCaseName);
 
 // Convolution/UnaryElementwiseArithmetic/BinaryElementwiseArithmetic is not supported boolean elemnt type
@@ -52,7 +52,7 @@ INSTANTIATE_TEST_SUITE_P(ov_plugin_numeric, CompileModelCacheTestBase,
                                  ::testing::ValuesIn(ovElemAnyNumericTypesTemplate),
                                  ::testing::ValuesIn(ovBatchSizesTemplate),
                                  ::testing::Values(targetDevice),
-                                 ::testing::Values(pluginConfig)),
+                                 ::testing::Values(ov::AnyMap({}))),
                          CompileModelCacheTestBase::getTestCaseName);
 
 // LSTMcell supported floating-point element type
@@ -62,7 +62,7 @@ INSTANTIATE_TEST_SUITE_P(ov_plugin_floating_point, CompileModelCacheTestBase,
                                  ::testing::ValuesIn(ovElemAnyFloatingPointTypesTemplate),
                                  ::testing::ValuesIn(ovBatchSizesTemplate),
                                  ::testing::Values(targetDevice),
-                                 ::testing::Values(pluginConfig)),
+                                 ::testing::Values(ov::AnyMap({}))),
                          CompileModelCacheTestBase::getTestCaseName);
 
 const std::vector<ov::AnyMap> default_properties = {
@@ -72,7 +72,7 @@ const std::vector<ov::AnyMap> default_properties = {
 INSTANTIATE_TEST_SUITE_P(ov_plugin, CompileModelCacheRuntimePropertiesTestBase,
                         ::testing::Combine(
                                 ::testing::Values(targetDevice),
-                                ::testing::ValuesIn(ov::test::conformance::generate_ov_configs(default_properties))),
+                                ::testing::ValuesIn(default_properties)),
                         CompileModelCacheRuntimePropertiesTestBase::getTestCaseName);
 
 } // namespace
