+++ src/tests/functional/plugin/conformance/test_runner/conformance_infra/src/skip_tests_config.cpp
@@ -20,8 +20,6 @@ const char *refCachePath = "";
 std::vector<std::string> IRFolderPaths = {};
 std::vector<std::string> disabledTests = {};
 
-ov::AnyMap pluginConfig = {};
-
 ShapeMode shapeMode = ov::test::conformance::ShapeMode::BOTH;
 
 } // namespace conformance
