+++ src/tests/functional/plugin/shared/src/main.cpp
@@ -7,14 +7,12 @@
 #include "functional_test_utils/summary/environment.hpp"
 #include "functional_test_utils/summary/op_summary.hpp"
 #include "functional_test_utils/skip_tests_config.hpp"
-#include "base/ov_behavior_test_utils.hpp"
 
 #include "set_device_name.hpp"
 
-bool ov::test::behavior::APIBaseTest::is_print_k = false;
-
 int main(int argc, char *argv[]) {
     ov::test::utils::disable_tests_skipping = false;
+    ov::test::utils::is_print_rel_influence_coef = true;
     bool print_custom_help = false;
     std::string outputFolderPath(".");
     for (int i = 0; i < argc; ++i) {
